+++ tests/urls.py
@@ -1,10 +1,9 @@
 from rest_framework.routers import DefaultRouter
 
-from tests.fixture_app.viewset import CriteriaViewSet, GenreCriteriaViewSet, TrackViewSet
+from tests.fixture_app.viewset import CriteriaViewSet, TrackViewSet
 
 router = DefaultRouter()
 router.register(r"criteria", CriteriaViewSet, basename="criteria")
-router.register(r"genre-criteria", GenreCriteriaViewSet, basename="genre-criteria")
 router.register(r"tracks", TrackViewSet, basename="tracks")
 
 urlpatterns = router.urls